fabric_adls_lakehouse_poc/requirements.txt
azure-identity>=1.15.0
azure-storage-file-datalake>=12.14.0
requests>=2.31.0
tenacity>=8.2.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
tqdm>=4.66.0
